A whole lotta nothing done
I tried to fix that stupid while loop in order to incorporate the distance sensor but it is above my pay grade. It it not moving on from the DROP_SAMPLE_S State. I tried putting a timer into the code to force it to move on but to no avail no success.

However there is basic code for Schuys new toy to work, the servo values need to be fine tuned, but the telemetry data and buttons are in there.

-A tired teenager signing off

Co-Authored-By: BobtheShrimp <116126182+BobtheShrimp@users.noreply.github.com>