Auto is Auto
So the Distance Sensor wasn't actually being used until the end of the meeting-not sure if it works correctly. It is reading but I am not sure if the motors are being turned on. The Turn Speed has been adjusted for the higher speed to be used at 45 degrees difference and up. Small changes have been made to targets