WE....WE SCORED 3 SPECIMEN IN A ROW
Little tweaks have been made to targets to help with overshoot as well as changing what the targets profile is.

We no longer use the distance sensor to move to the wall and instead use a timer to drive towards the wall.

Fixed the claw problems by using a if the specimen motor is not busy and then having the claws move down

Co-Authored-By: BobtheShrimp <116126182+BobtheShrimp@users.noreply.github.com>