OMG WE ARE SO IN THE GAME
We are now consitently scoring three specimen onto the bar

We took Prestons suggestions into account

We made our own gyroturn, so during that one slow turn it has been sped up and more accurate. We also added where at the end the robot is a bat out of hell and parks.

IT IS AMAZING AND GLORIOUS

-one small change done to left autonomous but that is not my baby yet...

Co-Authored-By: BobtheShrimp <116126182+BobtheShrimp@users.noreply.github.com>
Co-Authored-By: PlanB5309 <43391798+PlanB5309@users.noreply.github.com>