Left Odometery- Progress?
Changed the orientation of the starting position so had to create new targets accordingly. The robot drives away from the wall, lines up on the basket, drives forward, scores, and then lines up on the samples, drives forward and picks one up

That's all for tonight but once those targets are locked in they can be used again for the most part.

Co-Authored-By: BobtheShrimp <116126182+BobtheShrimp@users.noreply.github.com>