Left Odometery - Scoring three (almost) samples
Added new Targets to go with the states,  its a lot of repeating in regards. Murked a lot of the score sample code- a lot of it was repeated in earlier instances of code, it gives us more control of it. A little more tweaking and we will consistently have it.

Co-Authored-By: BobtheShrimp <116126182+BobtheShrimp@users.noreply.github.com>