Autonomus consistency?
For Left Odometery, we upped the minimum speeds by .5, now it runs just a little bit faster. We also upped the maximum angle different to 25 for when we are at the basket, and instead do a gyroturn once we're there-allowing us to be much more conisitent in getting closer to the basket. For the weird pause, when the arm is 75% to it's position we go ahead and move on.

Right Odometery
made the turn near submersible a little less for the third specimen, I backed up the sample target so it isn't running into it anymore.

Fixed the unblock arm jazzerisize as well.

Co-Authored-By: BobtheShrimp <116126182+BobtheShrimp@users.noreply.github.com>