Auto changes >:) A WHOLE NEW PROGRAM
We made the waiting change for picking up the specimen happen. GIRLBOSS.
Then we got a wild hair and started writing a whole new auto program to pick up 4 specimen on the right side cause we are girlbosses. We've gotten as far as moving two samples over to the observation zone. Timing and all that other jazz needs to work.